Fixing compile warning on amd64.

+++ idlestat.c
@@ -429,7 +429,7 @@ static struct cpuidle_datas *idlestat_load(const char *path)
 	fclose(f);
 
 	fprintf(stderr, "Log is %lf secs long with %d events\n",
-		end - begin, count);
+		end - begin, (int)count);
 
 	return datas;
 }
